feat: add translations to clearChatButton
