+++ src/index.ts
@@ -6,7 +6,7 @@ export { default as MentionButton } from './components/Mentions/MentionButton';
 
 // Admin tools to export and clear the chat
 // should be used in the item settings
-export { default as ClearChatButton } from './components/AdminTools/ClearChat';
+export { default as ClearChatButton } from './components/AdminTools/ClearChatButton';
 export { default as DownloadChatButton } from './components/AdminTools/ExportChat';
 
 export type { Member, ChatMessage, AvatarHookType } from './types';
